Fixed double promotion warning

+++ src/device/CalibrationHandler.cpp
@@ -293,7 +293,7 @@ float CalibrationHandler::getFov(CameraBoardSocket cameraId, bool useSpec) {
     int width, height;
     std::tie(intrinsics, width, height) = CalibrationHandler::getDefaultIntrinsics(cameraId);
     auto focalLength = intrinsics[0][0];
-    return 2 * 180 / M_PI * atan(width * 0.5f / focalLength);
+    return 2 * 180 / ((float) M_PI) * std::atan(width * 0.5f / focalLength);
 }
 
 uint8_t CalibrationHandler::getLensPosition(CameraBoardSocket cameraId) {
